+++ dynamodel_derive/src/lib.rs
@@ -1,9 +1,11 @@
 mod case;
 mod impls;
+mod types;
 mod utils;
 
 use case::RenameRule;
 use darling::{util::WithOriginal, FromDeriveInput};
+use impls::{enums, structs};
 use proc_macro::TokenStream;
 use proc_macro_error::{abort, proc_macro_error};
 use quote::quote;
@@ -17,7 +19,7 @@ use syn::{parse_macro_input, spanned::Spanned, DeriveInput};
 struct TargetStruct {
     ident: syn::Ident,
     generics: syn::Generics,
-    data: darling::ast::Data<utils::Variant, WithOriginal<utils::Field, syn::Field>>,
+    data: darling::ast::Data<types::Variant, WithOriginal<types::Field, syn::Field>>,
     rename_all: Option<syn::Lit>,
     extra: Option<darling::Result<syn::Path>>,
     tag: Option<String>,
@@ -41,61 +43,35 @@ impl TargetStruct {
         Ok(self)
     }
 
-    fn token_stream_struct(self) -> TokenStream {
-        let ident = self.ident;
-        let rename_rule = self
-            .rename_all
+    fn rename_rule(&self) -> RenameRule {
+        self.rename_all
             .as_ref()
             .map(RenameRule::from_lit)
-            .unwrap_or_default();
+            .unwrap_or_default()
+    }
+
+    fn token_stream_struct(self) -> TokenStream {
+        let ident = &self.ident;
         let (imp, ty, whr) = self.generics.split_for_impl();
-        let fields = self.data.take_struct().unwrap().fields;
+        let rename_rule = self.rename_rule();
 
-        let setters = fields.iter().filter_map(|f| {
-            if f.parsed.skip_into.is_some_and(|v| v) {
-                None
-            } else {
-                Some(impls::structs::setter_token(&f.parsed, &rename_rule))
-            }
-        });
+        let fields: Vec<types::Field> = self
+            .data
+            .take_struct()
+            .unwrap()
+            .fields
+            .into_iter()
+            .map(|f| f.parsed)
+            .collect();
 
-        let getters = fields
-            .iter()
-            .map(|f| impls::structs::getter_token(&f.parsed, &rename_rule));
+        let into_hashmap = structs::into_hashmap(&self.extra, &fields, &rename_rule);
 
-        let set_extra = if self.extra.is_some() {
-            quote! { item.extend(extra_item); }
-        } else {
-            quote!()
-        };
-
-        let init_extra = match self.extra {
-            Some(Ok(path)) => {
-                quote! {
-                    let extra_item: Self = #path(&value);
-                }
-            }
-            Some(Err(err)) => {
-                abort! {
-                    err.span(), "Invalid attribute #[dynamodel(extra = ...)]";
-                    note = "Invalid argument for `extra` attribute. Only paths are allowed.";
-                    help = "Try formating the argument like `path::to::function` or `\"path::to::function\"`";
-                }
-            }
-            None => quote!(),
-        };
+        let try_from_hashmap = structs::try_from_hashmap(&fields, &rename_rule);
 
         quote! {
             impl #imp ::std::convert::From<#ident #ty> for ::std::collections::HashMap<String, ::aws_sdk_dynamodb::types::AttributeValue> #whr {
                 fn from(value: #ident #ty) -> Self {
-                    #init_extra
-
-                    let mut item: Self = ::std::collections::HashMap::new();
-                    #(#setters)*
-
-                    #set_extra
-
-                    item
+                    #into_hashmap
                 }
             }
 
@@ -103,27 +79,21 @@ impl TargetStruct {
                 type Error = ::dynamodel::ConvertError;
 
                 fn try_from(item: ::std::collections::HashMap<String, ::aws_sdk_dynamodb::types::AttributeValue>) -> Result<Self, Self::Error> {
-                    Ok(Self {
-                        #(#getters,)*
-                    })
+                    #try_from_hashmap
                 }
             }
         }.into()
     }
 
     fn token_stream_enum(self) -> TokenStream {
-        let ident = self.ident;
-        let rename_rule = self
-            .rename_all
-            .as_ref()
-            .map(RenameRule::from_lit)
-            .unwrap_or_default();
+        let ident = &self.ident;
         let (imp, ty, whr) = self.generics.split_for_impl();
+        let rename_rule = self.rename_rule();
         let variants = self.data.take_enum().unwrap();
 
-        let into_hashmap = impls::enums::into_hashmap(&ident, &self.tag, &rename_rule, &variants);
+        let into_hashmap = enums::into_hashmap(ident, &self.tag, &variants, &rename_rule);
 
-        let try_from_hashmap = impls::enums::try_from_hashmap(&self.tag, &rename_rule, &variants);
+        let try_from_hashmap = enums::try_from_hashmap(&self.tag, &variants, &rename_rule);
 
         quote! {
             impl #imp ::std::convert::From<#ident #ty> for ::std::collections::HashMap<String, ::aws_sdk_dynamodb::types::AttributeValue> #whr {
